fix: rename blog to essay
