+++ src/App.js
@@ -1,14 +1,14 @@
 import "./App.css";
 import { Route, Routes } from "react-router-dom";
 import Home from "./routes/home/Home";
-import Blog from "./routes/blog/Blog";
+import Essay from "./routes/essay/Essay";
 import ErrorPage from "./error-page";
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/blog/:blogId" element={<Blog />} />
+      <Route path="/essay/:essayId" element={<Essay />} />
       <Route path="*" element={<ErrorPage />} />
     </Routes>
   );
