+++ src/essay/index.json
@@ -1,6 +1,6 @@
 [
   {
-    "location": "blog/00-test",
+    "location": "essay/00-test",
     "author": "Sahil Bondre",
     "date": "Week 52, 27th December 2022",
     "title": "Heading: Lorem Ipsum",
